requirements.txt
fastapi
psycopg2-binary
pydantic
SQLAlchemy
uvicorn
requests
bs4
python-decouple
pytest